rag/requirements.txt
fastapi
uvicorn[standard]
GitPython
boto3
qdrant-client
sentence-transformers
torch
transformers
networkx
pydantic
requests
faiss-cpu
minio